back/requirements.txt
django
mysqlclient
pymysql
djangorestframework
django-cors-headers
